+++ R/autoplot-response-spct.r
@@ -565,6 +565,10 @@ autoplot.response_spct <-
                 default = c("boxes", "labels", "summaries", "colour.guide", "peaks"))
     annotations <- decode_annotations(annotations,
                                       annotations.default)
+    # avoid warning in 'photobiology' (== 0.10.10)
+    if (is.character(norm) && norm == "update" && !is_normalized(object)) {
+      norm <- "skip"
+    }
     # normalization needs to be redone if unit.out has changed
     object <- photobiology::normalize(x = object,
                                       range = range,