+++ R/autoplot-source-spct.r
@@ -655,6 +655,10 @@ autoplot.source_spct <-
                 default = c("boxes", "labels", "summaries", "colour.guide", "peaks"))
     annotations <- decode_annotations(annotations,
                                       annotations.default)
+    # avoid warning in 'photobiology' (== 0.10.10)
+    if (is.character(norm) && norm == "update" && !is_normalized(object)) {
+      norm <- "skip"
+    }
     # normalization needs to be redone if unit.out has changed
     object <- normalize(x = object,
                         range = range,